+++ CHANGES.md
@@ -17,6 +17,7 @@
 - libwebrtc を 64.5.0 に上げた
 - deprecated warning を潰した
 - Signaling connect 時に client offer の SDP を載せた
+- Kotlin 1.2.30 に上げた
 
 ### FIX